refactor: consolidate two related plugins
